cleaned up some code and added "user_matched"
we're not gonna use "user_changed" in the daemon, so i removed that
code. what we are going to use on the other hand is the new
"user_matched". this is a reference to a track, and it's meant that the
user "connects" a file to a track using this field. the daemon will then
use the track data rather than the file data.