fixed bug in PostgreSQL
argh, was too eager earlier. can't both INSERT and UPDATE files, when we
move a file that means we'll INSERT the new file and then try to UPDATE
the old file to have the same filename as the file we just added (which
ofcourse will give us a duplicate key). this bug should be fixed now