made matched files more visible in album page
instead of a checkbox there will now be "OK" for files already matched.
this made is significantly easier to tell why certain albums weren't
saved.
also realized that i need a "saved" field in file table to be able to
tell which files are actually matched & saved and which are just
manually matched by the user