removed stat()'ing old files
instead of stat()'ing every single file in the archive to see if they're
still there we'll mark every file in the database and remove the mark
when we load the file from the database. at the end, files that are
still marked are gone and we can remove those.
also removed matches for files in a group before matching that group
with albums.