added "progress" to table "locutus"
this field will be updated as locutus is matching files. the value range
from 0.0 to 1.0. we can also use this field to guesstimate how much time
is remaining. the field is updated for each group, and since group size
varies it won't be extremely accurate, but should be good enough