modified header.tmpl slightly
moved "album matching" to the left and made "artists", "albums",
"tracks" and "files" grouped together as "cache". people are not really
supposed to mock around in the cache, it's just there for "easy access".
we're going to need a new page "unmatched files". this page will show
files that haven't got an entry in the "match" table. it's highly likely
that these files got very poor metadata or that they don't exist in
musicbrainz