*huge* speedup on album_matching page
the view used for the album_matching page was not properly formed. it
used a "LEFT JOIN" where it really just should be a "JOIN" which caused
it to return data about every single album & track in the database. as
the cache grows, that view would also consume a lot more time. since
we're not interested in albums that got no unmatched files we don't need
that "LEFT JOIN", and changing it to just a "JOIN" made the query
execute in milliseconds instead of seconds.
also added "album score" (which is just the sum of the track match
scores divided by tracks on album)