the huge terminology rewrite
most likely i broke at least one thing in this commit. if i'm to guess,
i'd say 42 things broke.
i've pretty much removed most references to "match", a comparison
between a file and a track is now called "Comparison". when we "connect"
a file to a track this will now be a "match". thus, the "match" table is
renamed "comparison", and more interestingly, "matched" in file table is
now called "track_id" as that's what it really is.
also, since we got a lot of files with filenames in various character
encodings i've changed the "filename" field from "varchar" to "bytea".
i'm not sure if it'll work and it really is a bit messy, but i believe
the solution is better than creating renaming files every now & then.