working on new filenamer
current filenamer is quite limited. i'm working on one that handles
functions as well as "variables" (stuff taken from metadata) and
"static" (stuff that's neither functions nor variables, like "/")
entries.
not yet done with this, it's a large task (guess i should've made an own
branch for this, but oh well, i'll just push to the "backup" repo for
the time being)