bugfix, saved albums it shouldn't in rare cases
i introduced a bug when i added "only save if every file in group match
something", causing it to save full albums even if there is a non-full
album with better score. the bug was a bit nasty as it could save short
albums (singles, EP) with erronous data when the metadata_min_score
setting was low, although it's unlikely a lot of files were affected by
this bug.