fixed bug #307898 for now
currently it'll only make sure the values it operates with is within the
correct range, but it may be an idea to improve this later. for example,
nothing prevents the user from setting "metadata_min_score" to 0.0,
which will make it save most likely every single file, regardless of how
horribly wrong it is. should we let users stab themselves in the eye
with a fork?