added "%ext%", "$bytes_left()" & "$bytes_right()"
we used to append extension after we built up the filename, but this was
a bad idea. for one, user might want to force lower (or upper) case
extension, but more critical: it made it harder to limit the filename on
filesystems where the there's a limit to the filename (and that's about
every common filesystem out there)
we can't use $left() and $right() for limiting filename because that
limits on characters, not bytes. these two new functions limits on bytes
safely (ie. if your filename ends up with 256 bytes where the last
character is a two-byte unicode character then it'll remove the
character (and not the last byte in the character) and end up with a 254
byte long filename).