grouping on album again
we used to group on musicbrainz_albumid or path to directory if the
albumid didn't exist. however, we got a problem when groups become large
and contained files from several albums.
for example:
folder with 1500 files, most from different albums. looks up album for
first file, match every file with every track on album, few matches.
look up album for second file, match every file with every track, few
matches, etc. if we loaded 500 albums with an average of 20 tracks that
means 15000000 comparisons between files & tracks, and with each
comparison eating 48 bytes then we're nearly up to 700mb memory needed.
the cpu needed to fuzzy match also takes its toll on the cpu.
this is a difficult problem, and i haven't come up with a better
solution than trying to split the files up in smaller groups. the
problem with splitting up the groups is that it may decrease matched
albums as we may have files that match the same album, but with
different metadata in the "album" field. however, the new
"combine_groups" setting helps on this.