probably fixed issue with memory not being cleared
this one was a bit tricky. in an attempt to reduce amount of malloc() &
free() i reused maps & vectors. however, these ate up a large amount of
memory, while it appeared like it wasn't the maps/vectors eating the
memory, but rather the data in those structures.
it is however changed to actually delete the structures now, so in
theory it should not hold on to memory anymore (not tested).