improved genre handling
previously we've looked up genre for every single file we've sorted.
given that it takes at least a second looking up genre per file (in some
cases it may even take 2 seconds) this change will significantly reduce
the amount of time locutus require:
- only lookup genre when we move a file or if the file's genre is set to
  "".
this still means that files where no genre can be found will be looked
up repeatedly, but hopefully this only applies to a handful of files and
shouldn't be a large issue.
also made it write "moving file" to the debug log when it's actually
moving a file, and not every time it saves a file.